MBS-10761: Disallow YouTube links at wrong levels (#1460)
Users keep wrongly adding channel links to recordings and video links
to artists. This commit disallows that, since there's no case where a
video link is a channel nor a channel is a video link.

For video links we already standardise them to the "watch?v=" syntax
so it should be safe to disallow all others. For channel links YouTube
seems to have a ton of different options, all of them valid, so this
just checks it's a YouTube video and not a watch video.

Releases are sometimes officially put up on YouTube as playlists.
As such, it makes sense to allow playlist links for them.