MBS-10802: Improve alignment of Reorder Relationships table
The Relationship header being on the right is consistent with how
we usually do edit headers, but not with just how long these lines
often are, so moving it to the left. Similarly, there's no reason
for the order columns (just numbers) to take so much space, so this
makes them take only the space they need.