MBS-10732: Add collaborator validation to collection form
Up until now a user could type in an editor name in the collaborators
field for a collection and try to submit without selecting a result,
leading to an ISE.

The error field is special-cased to remove the margin because otherwise
a margin of over 160px makes it look absurd.

Also, inputName for the autocomplete was incorrect, so this
fixes that.