MBS-10811: Convert historic Edit Relationship edit to React
Changes from the TT:

Added the edit classes to be more consistent and improve the display of the data.

Includes a hack to allow passing the different link phrases
with the link types.

If we managed to add the attributes properly to these edits
at some point we could remove the hack.