MBS-10138: Fix guess case for titles finishing in "f" (#1495)
Guess case tries to find whether a title includes "f/" or "f."
(for featuring artists) by finding a solitary F, and then checking
if the next "word" is "/" or ".". This usually works, but since it does
it by trying to match the next word, we need to check whether
said word actually exists at all before attempting to match.
Otherwise, an F at the end of a title will lead to an attempted
match on null (since there's no next word) and Guess Case crashing.