MBS-9340: Only allow mul and zxx as the only work language
There's no case where it makes sense to have [Multiple languages]
or [No lyrics] and a second, different language as a work language
at the same time.

There's already a limitation on the JS layer that makes it so no more
languages can be added if one of the two is selected, but that
doesn't affect any languages already selected before them.

I chose not to further enforce this on the JS level because the
form verification does all that is needed while making sure the selected
languages do not get lost/deselected, which would seem annoying
if the user selected one of these two by mistake.