MBS-10187: Fail edits that would result in a gendered group
An edit that merges a group into an artist of no type set, but with
a gender set will currently crash at the DB level. This adds an error
and fails the edit - we could also choose to silently drop the type, but
this seems better for avoiding dubious / dodgy merges.

NB: this will also cause an error in a situation where the edit would
possibly have applied earlier - if both group *and* person artists
are being merged into the no-type, gendered artist. That situation
seems even more dodgy though, so I think that's actually a positive
change.