MBS-9931: Fail gracefully when trying to remove example rel
If someone tries to remove an example relationship, the edit gets
stuck. We don't want people to actually remove these themselves,
so this fails the edit and asks them to contact us.

Eventually, we might want to block entering these removals
completely, but that's much harder and given the situation is
pretty rare, this seems like a good first step.