MBS-10834: Account flag to disable writing edit notes
The editing_disabled flag blocks a user from editing, but they can
still leave edit notes on edits. This is fine in cases where the editor
just makes bad edits (whether by lack of care or knowledge),
but in some cases the problem is one of attitude when interacting
with the community. In that situation, just blocking editing won't
stop the user from leaving insulting / demeaning edit notes.
This is the goal of the new flag. In some cases, we might want to set
this *without* actually disabling editing at all, especially if a user
keeps getting into a comment war and just needs to stop commenting
for a while.