MBS-10842: Remove report link from deleted editors
There seems to be no reason to allow reporting deleted editors,
since they already deleted their account. I talked to Freso and he
agreed. If there's a particular problem they still want us to look at
they can contact us at support.

This changes the /report page to say that - I could have blocked it
if the user has been deleted, but it feels that if somebody makes
the effort of trying to get to /report without the link, we might
as well explain the issue.