MBS-10841: Add "Guess case" per medium
Sometimes, for example, one medium will have English titles and
another titles in a language that requires sentence guess case,
or most discs in a boxset are correct (and might become worse
with automatic guesses) but one or two need guess case.
This add an extra Guess case button under each medium
(there's already plenty of buttons there so it fits well) that affects
only the tracks for that medium plus the medium title.