MBS-10840: Fix key names in English guess case (#1525)
Guess Case currently does a pretty bad job with classical key names:
they should be "in B-flat major" or whatnot as per the guidelines,
but the guess case code just capitalizes them as any other words.
This is normally annoying, and sometimes kinda ridiculous
("Sonata in a Minor" is a quite  different from "Sonata in A minor"!).
The code only runs for key names preceded by "in", to avoid
false positives ("Signed to a Major" or whatnot). I'm sure there
are still some cases where this will give a false positive, but I'm
pretty sure they will be much less common than the classical releases
it currently mangles.