MBS-10845: User lookup tools for account admins
/admin/email-search:

Allows account admins to search users by email using a POSIX regular
expression. Since I didn't bother paging the results, it's limited to
returning 100 users (in case someone were to accidentally enter a
pattern matching half the database) ordered by signup date.

/admin/ip-lookup/$hash:

Allows account admins to see a list of users accessing MusicBrainz from
a specific IP hash (an MD5 of their IP address). This is linked from
users' profile pages, which now display a list of hashed IPs "belonging"
to a user to account admins. (The point of the hash isn't to be
cryptographically secure, since admins can find this data in the server
logs anyway; it's just to obscure the possible geographic location of
the user, since it doesn't seem relevant in most cases.)

Redis keys mapping a user ID to a list of IP hashes and vice versa (an
IP hash to a list of user IDs) have an expiration of 6 months, but this
expiration is refreshed whenever a relevant user or IP accesses the
site.