MBS-10499: Alert for entity in danger of removal once add edit passes
Sometimes users (especially new ones) add an entity as a votable
edit (by checking the relevant checkbox) and are then not notified
that the entity is in risk of removal, since their add edit means
there's a pending edit keeping the entity around. Unless they have
added more data, though, the entity will be removed soon after the
one pending edit passes.

This checks if there's only one pending edit and that is the only
thing keeping the entity from being eligible for cleanup. If that's
the case, it then notifies the user with a banner.