MBS-10469: Add 10 minute delay to shown CAA on front page
Currently a lot of cover art on the front page is not shown since it
hasn’t been processed by the IA yet, so visitors see "Image not
available yet" messages instead. The proper solution to fix this would
be to not show images on the front page until they’re actually
processed, but I am not currently aware of a way to do this easily and
reliably (there are database columns for storing sizes of the
thumbnails, but they don’t seem to be populated, otherwise they would
possibly be a good fit).

This is mostly meant as a temporary workaround to lessen the problem. It
won’t remove the problem entirely, since sometimes cover arts take more
than 10 minutes to generate, but having a 3+ hour long wait also seems
excessive.