Fix format_editnote running on localized note URL
_localize_text runs recursively on any variable substitution arguments,
in case those are also localized (see Edit::Release::Merge for an
example). This can break if they aren't localized, because
format_editnote is run as a fallback in that case. It's unlikely we'd
ever be passing text as a substitution argument that's intended to use
edit note syntax, which is something for text entered by users, not
pre-defined server messages. So we now only run format_editnote if this
is a top-level call to _localize_text (meaning none of the edit note is
localized).