MBS-10187: "Merge artists" edit is stuck
Artist merges can get stuck if their merged types and genders conflict:
groups can't have a gender.

We now detect this case and drop any conflicting information. This is
done as opposed to failing the edit because it's less annoying to the
user. ModBot will leave a note explaining what information was lost.