MBS-10916: Display AcousticBrainz info in the Recording sidebar
Add AcousticBrainz information in the sidebar for recordings: number of
submissions, BPM (if existing) and tonality (if existing and confidence
percentage > 50%).
The information is fetched through the AcousticBrainz API for the first
submission, using a React hook in order not to delay the display of the
rest of the page.