Add basic Modal and Popover components
These will be used in the release/relationship editors, and for other
miscellaneous components like the guess case options and artist credit
editor.

Most of the complexity here comes from keeping the components accessible
by managing focus: either trapping it inside the dialog for modals, or
returning focus to the opening button once the modal or popover is
closed. (For comparison, react-bootstrap provides some ways to handle
this sort of thing for modals, but not popovers.)

A test fixture for the modals/popovers has been added at
root/static/scripts/tests/dialog.html. Currently this is only useful for
manual testing. You must run compile_resources.sh with the 'tests'
option before loading this page.