MBS-10901: Fix ISE when historic EditRelationship data is broken
We seem to have some cases where historic EditRelationship edits
do not have "new" for some of the "old" relationships in the old-new
pairs. This makes no sense, so it's probably just data corruption
caused during the NGS migration - we can't do much about it,
so this displays those old rels unchanged (diff with themselves),
which is the least neurotic of the possibilities, and adds a
warning notice so that the user will know what to expect.