MBS-10908: Stop recording-of rel edits showing for all recordings of work (#1574)
While it's generally useful to show relationship changes for a work
on recordings and releases connected to the work (MBS-6532),
for example if writers are being added or changed, there's one
very common use case (recording-work "recording of" rels)
where the edits are just noise on other recording entries. The fact
that other, unrelated recording is or isn't a recording of this work
should never affect the other recordings of the work in any way.
This makes it so the edit is only added to the edit histories of
recordings and releases connected to the work if the relationship type
is *not* "recording of". The recording and release(s) that are directly
connected to the edit will still be linked through the recording
side of the edit data.

Also, the previous implementation of MBS-6532 had an error in which
all changes made to any work also appeared on all artists who
ever recorded that work. This was unintentional, and this removes
it (if it gets explicitly asked for for non-"recording of" rels it can
be readded in a more intentional, not side-effecty way).