MBS-10926: Add 'copy date' button on relationship editor dialog
Add a button to copy quickly the "begin date" content to the "end date"
fields. Since the dates fields are implemented differently in different
parts of MusicBrainz, this solution works only on the Relationship
Editor, which is probably where this functionality is most needed.

The button will be displayed on the same line as and after the begin
date fields, which makes the most sense for a keyboard workflow (Tab key
between begin date fields, then Tab Enter to copy or Tab Tab to access
the end date fields).