Cache the number of hits in query_to_list_limited
Running the result count(*) of a paginated query for every page can be
expensive, as it naturally has to first retrieve all results from the DB
in order to run the count(*), then throw away any results outside of the
LIMIT+OFFSET (though it's true that it has to do this for rows up until
OFFSET anyway). This commit allows query_to_list_limited to optionally
cache the number of hits for a fixed amount of time (currently 30
minutes). This can make sense for lists that don't have their number of
pages change often, like releases by an artist or label. Less so for
lists of edits to be voted on.