MBS-10932: Match Instagram videos /p/ and /tv/
Until then, Instagram photos were disallowed to be linked to an entity,
but videos had not been even considered.

This patch allows linking an MB recording or release to an Ig video.

Since /p/ and /tv/ are interchangeable in Instagram URLs, and since the
default format used to link to a video is /p/ on an Instagram profile,
this patch also normalizes /tv/ URLs to /p/ instead.