MBS-10943: Fix loading attribute on Relationship::Delete
Always creating a new LinkAttributeType means we're missing
data that is necessary, for example, to decide whether to also
convert text_value to JSON, causing the bug here. This change
ensures that we use the actual attribute as loaded from the DB if
possible, in the same way we already do on Relationship::Create.