MBS-10956: Load ratings everywhere we also load user ratings
This is a generalization of MBS-10937. For some reason, a fair amount
of places where we load user ratings are missing a load_meta call,
so the average community rating is never loaded there (and the
rating stars will thus remain grayed out). This adds the load_meta calls
everywhere we already load user ratings for display.