MBS-10965: Convert Add Disc ID edit to React
The original TT code checked !edit.preview OR edit.display_data.medium
to decide whether to display the Medium section of the edit data.
That means that it would always be displayed if there's a medium though,
and if there isn't then there's nothing to display anyway, so I changed
it to just a medium check.