MBS-10973: Convert Add Relationship Attribute edit to React
This moves stuff that was being read from edit.data into display_data.
The parent header was shown in TT even if there was no parent;
changed it to only display if there's a parent attribute type.
This translates the parent type name (since it's not being added
or changed by the edit) but the name of the new attribute is left
untranslated since it's a free text field being added by the edit.