MBS-10923: Fix inconsistent ordering in find_by_collection (#1591)
`DISTINCT ON (release.id)` is used here because the joins create
"duplicate" rows for every release. With DISTINCT ON, only the first row
from each set is kept, so the ORDER BY in the inner query is important.
If `$order_by` is say `date_year, date_month, date_day, name COLLATE
musicbrainz`, then after sorting by id we also want it to consistently
pick the row with the earliest year, then month, etc. (There will be a
different row for each release event.)

In the outer query, we also want to make sure that releases that happen
to have the same release date and name (or whatever the configured sort
order is) are next ordered by id, otherwise the sort is undefined.

One gotcha here is that `$order_by` references column aliases that can
be used in the outer query, but not the inner one. We have to return a
separate `$inner_order_by` with explicit column references where needed.
This is icky but I couldn't come up with anything better. One solution
might be to replace all the joins with semi-joins, which would also
remove the need for DISTINCT ON, but I haven't evaluated the performance
of that.

It's not really feasible to write a test for this that fails
consistently, since the order is going to depend on the order PG
retrieves things, and that's unlikely to change in a very small test DB.