MBS-10986: Convert cannot_approve to React
This was trying to call auto_edit_for_editor on the TT, but that was
removed ages ago, and by now is equivalent to editor_may_approve.
Given that the check to show this page at all already implies
editor_may_approve is false, I just removed the "reason unknown" line.
We were checking for status === 8 for "edit has been cancelled"
but that's actually "TOBEDELETED", a deprecated transition status.
Changed that to actually check for DELETED instead.