MBS-10999: Make adding first IPI/ISNI an auto-edit
We generally allow auto-edits for everyone if data is being added
where it didn't earlier exist. For some reason though we didn't
make that check for IPIs and ISNIs. It makes sense to make the edit
go into voting for cases where an IPI or ISNI already exists, but
otherwise we should act in the same way as we do for other data.

We check whether an ISNI / IPI is already in use elsewhere,
and block autoediting in that case. Whether the edit was otherwise
autoeditable or not, ModBot will leave a note specifying that
the IPI or ISNI is already in use in x artists / labels, and
linking to a search to find out which (actually building the
list for display on the edit turned out to be almost impossible
with localized_note, so this is the best compromise I could find)