MBS-10992: Disable ended URL links
If a URL relationship has an "ended" flag (e.g. an artist's homepage
is gone), this displays it as just text (a "disabled link").

Especially significant in cases where the URL has been taken over
by a cyber-squatter/scammer or some other disreputable site
that we do not want to provide links (and possible page-rank) to.