MBS-11004: Support label lyrics links
This is needed for laboiteauxparoles.com.

While this will allow label lyrics links for all lyric sites
not ran through validation, we also do not check that those
should allow all three of artist, RG and work. As such, it seems
fine to allow this for now until we introduce proper validation
for all of them.