MBS-11014: Add context to DuplicateViolation on rels editor
While this will be fixed more properly in the React conversion
(where the message will be shown by the affected relationship,
so we can revert to the original string), in the meantime
this string gives a bit more useful information than just
"This relationship already exists" at the bottom of the
relationship editor.

The original string, which should be re-added on the React
conversion, is still in use in several other places so the
translation will not be lost.