Set es2020: true in .eslintrc.yaml
This at least fixes eslint complaining about the 'Map' and 'Set' globals
being undefined.

es6 was already enabled in script/check_eslint_rule, but not in the main
config, so the warnings still showed up in my editor. 'es6' is
technically enough to make Map and Set work, but we might as well use
the latest version since we're not limited to es6 by Babel or anything.