MBS-10994: Convert user/message.tt to React (#1618)
This reuses the existing StatusPage component, which works
perfectly for the use case (but should maybe be renamed by now).

email_sent.tt was effectively the same page but with the messages
defined on the TT instead of the Perl side. I moved them to the Perl
side so it can also use UserMessage.