MBS-10721: Improve track info in ArtistCreditBubble (#1444)
This solves several issues with the track names in ArtistCreditBubble:
1) Editing a new track's AC before adding a name showed [removed]
as the title. Now it will show [missing track name], with a tooltip
explaining that the user still needs to enter a track name.
2) Editing an existing track's AC after blanking its name showed
no title at all, making clicking through to the track impossible.
Now it will show [missing track name], but still link to the track.
3) Editing a new track's AC when a name had already been added
still set the "deleted" class on the track name and displayed
the "has been removed" tooltip. Now it doesn't show it as removed and
uses the caption "This track hasn’t been created yet."