Add invariant module
If Flow sees a function named `invariant` being called, it will behave
as if an exception was thrown inline where the passed condition isn't
true. The is useful for determining unreachable code and maintaining
refinements.

In other words, `invariant(foo, msg)` is equivalent to:

  if (!foo) throw new Error(msg);

in that Flow even refines `foo` as truthy below the condition in both
cases.