MBS-11052: Convert Edit Relationship Attribute edit to React
Changes:

We don't show a description line at all anymore if the description
is (unchangedly) empty.

Parent is no longer diffed by name, since the edit might change
the parent between two attributes of the same name. If the IDs
differ, the names are shown. In effect this probably won't happen
since attributes with the same name are rare to begin with, and the
display might still be confusing if a diff is shown between the same
names, but otherwise it wouldn't have shown a change happening at all.

We indicate the fields that have been explicitly left untranslated
with IntentionallyRawIcon.