Move hydrate props from attribute to <script>
Putting the props JSON into an attribute means that all of the quotes
(") in the JSON have to be encoded as `&quot;`, which inflates the
response size and makes the JSON impossible to read (making debugging
harder).

Instead, put the props JSON into a <script> tag immediately preceeding
the element to hydrate.