Transpile Jed, mutate-cow for IE11
node_modules/ is excluded from Babel processing, because it's very large
and almost everything under it is already suitably transpiled before
publishing.

Jed (or our fork of it), however, uses Object.assign which isn't
available in IE11. And mutate-cow uses boatloads of methods that'll
cause IE to scream. So these are added as exceptions (and they may not
be the only ones in the future).