MBS-11039: Don't use entity id as key for seriesItemNumbers (#1656)
An entity can appear more than once in a series with different
series numbers. While it is uncommon, it is supported and has real-life
use cases. With the current code, the number is overwritten and only
the latest number is displayed for all rows of the series table.

This changes seriesItemNumbers to be an array that is later accessed
based on the index of the react-table row (so, the entities array map).