MBS-11062: Link basic how-tos from the front page
The Quick Start section in the front page seems like a good place to
direct people to the most basic editing how-tos: artists and releases.
A general link to the how-to page is in the documentation menu, but
having these two somewhere that might be more directly accessible
to newcomers can't hurt.

While we have a lot more how tos, these two are the ones that seem
the most useful for new users, and from the bottom of those two
they can always find the links to the more advanced ones.