MBS-11085: Fix canHaveDiscID check
This was changed with cede01b1342c553ceb91e790f61eedcff26db5a8 -
I assume under the assumption that MB.formatsWithDiscIDs is an object.
It is not, though, it's just an array of IDs, so this was checking
whether the position in the array matching the format ID was null.
Changed to use includes, as it did before (just the native version).