MBS-8328: Add the non-writer artists back as own column
The previous commit removed these from the places meant specifically
for writers. In some lists though it can be interesting to have
non-writer artists listed too. This readds them to most general
"for display" lists (artist works, search results, ISWC index).
It does intentionally leave them off the IswcsWithManyWorks
report, because this data should be completely irrelevant
for the purposes of that report.