MBS-11061: Disallow MusicBrainz URLs in relationships (#1675)
I checked and there were only three more-or-less legitimate cases
of this (among a larger amount of stuff clearly pasted in error such
as links to the same entity the URL was added to). One of those
was a link to our wiki summit page as an "official website" link for
the Summit 2014, where some music got recorded (and arguably
that is not an official website anyway) and the other two were artists
using their own MusicBrainz user profile for a bio and then linking
to their user page as their biography page (which is something I
think we should disencourage anyway and feels very spammy,
while being no better than just adding the same text to an annotation).