MBS-11101: Fix series not using seriesOrderingFunc (#1695)
c5d54bd0ba5d178a4cb04c5bea83658ebaa4b278 changed this in a way that made
it so sorted(result()) wasn't ran at all on series with a seriesType but
without a seriesOrderingFunc. Due to this, no relationships were
displayed at all on such a series' edit page.