MBS-11107: Support nameVariation for other entity types
We can use relationship credits with most entity types. As such,
it makes sense to supply the same hover info + span class name
for them as we currently do for artists and recordings.

I special-cased URLs since I expect we don't really want to have
hover features for them if we print a pretty-name and not the URL.