MBS-11094: Don't block editing on pre-existing too early format
With other errors like this one, we allow the user to edit as long
as they don't actually introduce a problem that didn't previously
exist. This does the same, by not giving an error (but still displaying
a warning) if the too early format was already there. It will still
give an error if the user selects *another* too early format.