MBS-11127: Use cardinality for work-level-rels in WS/2
This avoids returning all unrelated recording-work relationships for
every release and recording requested with work-level-rels inc,
which in some cases was over 90% of the size of the full response
and which can always be retrieved from the work endpoint
if the data is actually needed.

The only other relationship currently not returned because
of cardinality is work-artist "artist was named after work",
which also seems irrelevant to release responses anyway.