MBS-11129: Display mediums as Medium # (Format)
The current way (CD 1, DVD 2) kinda makes it sound like
there should be a DVD 1.

For release tracklists, this just shows the medium position
at the beginning of the header, then shows only medium format
and name after the expand/collapse arrow.

Everywhere else, such as on edits, or on the relationship lists
under the release, Medium 1 (CD), Medium 2 (DVD) seems a lot
less confusing and not too repetitive / silly.