MBS-486: Add support for years BCE
We have had kinda semi-broken BCE support for decades, but we
have legitimate artists who were born and died BCE, and places
old enough that they were founded BCE, for example. This adds
proper support for BCE dates.

The dates are stored in standard astronomical format (1 BCE = 0)
but they are always shown to the user as the more intuitive version
without a year 0 on the website (including for entering and editing).
On the API, which is mostly meant for machine usage,
the astronomical format is retained.

Year 0 is blocked when editing since it's not a valid year in BCE
style. I added a more precise error when validating and in React
forms. I did not bother to implement the same error in TT, since
the date is rejected as invalid there anyway, which is still correct
(just slightly less clear).