MBS-10991: Convert user/subscriptions pages to React (#1617)
This was all separate pages, but it's not really necessary to keep
them separate. In fact, having only one file seems a lot cleaner to me.

Instead of passing collections to be displayed as "public_subscriptions"
this just passes them as "entities", like elsewhere, to avoid unneeded
differentiation.

layout.tt did a weird thing with N_ln functions in the summary
which seems to be completely unnecessary, so this does it more
normally with just ln.