MBS-11114: Allow track times to be unset (#1712)
0d45452121257d336c915f1b78f1114a21afe518 broke this by incorrectly
removing the "useless" isNaN check, which it turns out was not
"useless." The result of `unformatTrackLength` is null when the length
is empty, so the bug is that we'd bail out too early with the new check
and fail to set `this.length(newLength)`. The result is only NaN when
the input can't be parsed. A test is now added.