MBS-11036: Don't show Add release button when moving CD TOC (#1704)
This button is currently broken, but after some talk we decided
that it shouldn't be fixed, but dropped. There's no way to know
that moving (as opposed to additionally adding) a CD TOC
to a new release is correct, except for very specific cases.
It makes more sense to only allow moving the CD TOC to another
existing release.
If it has been added by mistake, it can either be removed
from the wrong one (and the new release can just be added based
on the CD TOC) or it can be moved after adding the release.