MBS-11146: Reviews tab for ratings + CB reviews
As shown on TV, er... talked about during Summit '20.

NB: the reviewable entities here are based on the ones
currently on CritiqueBrainz beta. The beta should be released
before this code is merged.

This creates a tab for ratings, which we didn't previously have
but makes sense in any case IMO.
For entities that can have reviews, the tab is named Reviews
and also displays the CB reviews. For entities which can have
ratings but no reviews (recording only at the moment)
the tab will just be called Ratings.
Places can have reviews but don't support ratings - for them
the ratings part of the tab is hidden.

We don't really want to display reviews without a body
(rating only reviews) from CB, but for now there's no way
to filter those out. This should be amended once that changes
(hopefully before release).

Flow is unhappy because it doesn't realize
entityProperties.ratings limits entities to RatableT and so on,
so I added two FlowIgnores for that.