MBS-11150: Don't send artist gid if none selected
This was sending artist.gid, but that button is meant to be used
"if you don't see the artist you are looking for", so submitting any
gid is contrary to the intent anyway.

Since artist got loaded in a foreach earlier, this was just
submitting the MBID of the last artist in the foreach loop.