MBS-11161: Fix error page display for actions using React
If a Catalyst action's response is rendered using React, it will have
set `$c->stash->{current_view}` to 'Node'. However, if an exception
occurs after this is set, we run the following bits of code inside
`finalize_error`:

    $c->stash->{template} = $timed_out ?
        'main/timeout.tt' : 'main/500.tt';
    ....
    $c->view->process($c);

Something's obviously wrong there: we're trying to display a .tt error
page, but processing it with View::Node. What happens is that the error
page is not displayed at all; the original action's React template is
rendered (which may display its own, different error from missing data).

Note that this code is also used for WS responses, so we don't want to
hardcode `$c->view('Default')` here either. Instead, I've added a
temporary fallback to 'Default' if `current_view` is 'Node'.