Move frame filtering to Devel::StackTrace call
I'd like to use the Devel::StackTrace object for displaying stack traces
on error pages, rather than using the one transmogrified for Sentry.
This requires performing the frame filtering (to remove irrelevant
items) beforehand. Luckily, Devel::StackTrace already allows for this.
(I'm not sure why it wasn't done this way to begin with.)