MBS-11160: ISE pages display empty stack traces
`$c->stash->{_stack_trace_info}` is now a structure resembling the
following:

    {
        $error_message => {
            as_string => '...',
            sentry_frames = [...],
        },
    }

This patch uses `as_string` to display stack trace info on ISE pages.
The standalone error message is used only if a stack trace isn't
available. (This can happen if `$c->error` is set directly without an
exception actually occurring.)

`$c->stash->{formatted_errors}` is used instead of `$c->stash->{errors}`
(which is just set to `$c->error`) for display. The 'Stack trace'
section has been removed and merged into the general Error(s) section,
because the displayable error strings will include the stack trace if
it's available.

We don't render the frames as an unordered list any longer. The stack
trace is simply displayed inside a <pre> tag. There's no need to format
the frames as before, because this information is for debugging purposes
and doesn't have to be structured in the markup. The new output even
includes more info and is easier to copy.